Add GPLv2 file
Simply stating "GPLv2" at the bottom of the readme isn't enough to
apply the license.

This PR is a start on properly applying the GPLv2 license to the
project code.

Much more remains to be done to apply the license correctly:

* Add a copyright & license statement to the top of all project files
(which are not already licensed by their own copyright holders)
* Update the README with any additional licensed files which are used
in the project (how about those images? what are their
copyright/licensing?)

Please see https://www.gnu.org/licenses/gpl-howto.html for complete
information on applying a GPL license to a project.